cheri: Manually apply fixes for FixPtrToULongCast warnings
Manually add fixes for the FixPtrToULongCast warnings where
clang-tidy isn't able to fix them automatically due to macro
expansions.

Explicitly ignore warnings in fs/btrfs/ because this file system
plays tricks with pointers that need a different fix.

Ignore some warnings that need a different fix.

+++ include/linux/compiler.h
@@ -149,7 +149,7 @@ void ftrace_likely_update(struct ftrace_likely_data *f, int val,
 #endif
 
 #ifndef RELOC_HIDE
-# define RELOC_HIDE(ptr, off) ((typeof(ptr))((unsigned long)(ptr) + (off)))
+# define RELOC_HIDE(ptr, off) ((typeof(ptr))((uintptr_t)(ptr) + (off)))
 #endif
 
 #define absolute_pointer(val)	RELOC_HIDE((void *)(val), 0)
